Add new safety issues to safety policy file feb 2024
Adds new safety entries for:
* GitPython
* Jingja2
and update the minimum versions where appropriate in
the dev-requirements/minimum-constraints files.

Add note about the version of Jinja2 which fixes that issue.

Update send2trash to version 1.8.2 since that is the minimum
version for the jupyter-server dependency.

Add changes for two new pylint issues in:
* statistics
* list of pylint ignores

Limit version of build package to > 1.1.0

* The build package 1.1.0 no longer provides the wheel package in its virtual
  env. pywbem currently depends on the wheel package for building.
  This change circumvents the issue by pinning buuild to <1.1.0, and adds
  a TODO to remove the dependency on build.

Increased Sphinx version and reduced to <py38

* Increased versions of Sphinx and its dependents to resolve compatibility
  issues with Jinja2 3.1.3 (which is needed to resolve safety issues).
  Reduced the environments in which Spinhx is installed to Py>=3.8, in order
  to remove too old Sphinx package versions.
  As part of that, increased the minimum versions of snowballstemmer, imagesize,
  and Babel.

Signed-off-by: Andreas Maier <andreas.r.maier@gmx.de>