Add fixes for new safety issues, Mar 24.
Modified safety policy dev-requirements, and minimum constraints for new
issues with:

1. cryptograpyh - New minimum 42.0.2 for python > 3.7
2. Jupyter Notebook - New Minimum 7.07 Python > 3.8

The new safety issues are as follows:

   Vulnerability ID: 65183
   Affected spec: >=7.0.0,<=7.0.6
   ADVISORY: CVE-2024-22420 describes a vulnerability in Jupyter
Notebook, where user interaction with a malicious notebook or
Markdown file
enables an attacker to access and act with the same permissions as
the user.
The flaw lies in the table of contents plugin. Jupyter Notebook
v7.0.7
includes a patch for this issue. Users can manually disable the
plugin as a
   workaround.
https://github.com/jupyterlab/jupyterlab/security/advisories/GHSA-4m77-cmpx-
   vjc4
   CVE-2024-22420
   For more information, please visit
   https://data.safetycli.com/v/65183/f17

   NOTES:
     New minimum version: 7.0.7

     7.0.2 python 3.8, 3.11
7.0.7, first version with documented python 3.12 support.
Previously
     the classiifier was python 3:;

-> Vulnerability found in notebook version 7.0.2
   Vulnerability ID: 65182
   Affected spec: >=7.0.0,<=7.0.6
   ADVISORY: CVE-2024-22421 is a vulnerability in Jupyter Notebook where
clicking a malicious link could expose Authorization and XSRFToken
tokens to
third parties in older jupyter-server versions. Patched versions
include
notebook above 7.0.7. Users are advised to upgrade jupyter-server to
version
2.7.2 or newer, which includes a fix for a redirect vulnerability. No
other
workaround has been identified.
https://github.com/jupyterlab/jupyterlab/secu
   rity/advisories/GHSA-44cc-43rp-5947
   CVE-2024-22421
   For more information, please visit
   https://data.safetycli.com/v/65182/f17

   Note: NEW MINIMUM 7.0.7  Also proposes jupyter-server >= 2.7.2