Add interthread queue to indication listener
Adds a queue so that indications received from the network are put
into the queue and a separate thread removes the indications from
the queue and sends the to the defined callback functions.

Includes the code to stop the queue thread.  Does not include code
to stop the http or https threads.