Fix issue where listener startup failure can cause extra exceptions.
Sets the callback thread to daemon so that in all cases the failure to
start the server stops this thread and creates a specific method
for the callback thread shutdown so that it is shutdown in case
of http/https server exception during startup.

This fixes issue #3157 where the failure of the http server
in pywbemcli startup causes the subprocess to hang because the callback
thread was already created and was not shutdown in the exception.