Modify documentation on cim types for CIMInt.
Modify the documentation on the CIMInt type definition to publicly
support both the positional and  keyword definition of the input
argument 'x' that defines the value of the CIMInt.

This differs from the definition of int() where the definition of the
value argument is only positional for Python >= 3.7.  This provides
compatibility for pre python 3.7 apps that use the CIM integer
types Uint*, Sint*.

One other minor fix where there were still several uses of long in the
documentation in pywbem. These were removed since effective with
Python 3.8, all of these types are now int.