Set the pywbem listener threads to daemon=False.
This forces the threads to be closed by the join statements for
each thread (and the processing of all received indications by
the callback thread) before closing the WBEMListener object.

No listener should ever exist after the WBEMListener object is
closed.

Changed a number of message outputs where we had inserged # in
front of integers in display in horrible integer to indicate that
this was a number.  Remove all of the # characters and clarified
the message texts slightly. This was mostly in the tests.